fixed unit tests

+++ UnitTests/BigDecimal_UnitTests.cs
@@ -16,7 +16,7 @@ public class BigDecimal_UnitTests
     [InlineData(10.0, true, 1.0, 1)]
     [InlineData(0.01, true, 1.0, -2)]
     [InlineData(-3254.214, false, 3.254214, 3)]
-    public void TestConstructorDouble(double d, bool sign, double mantissa, int exponent)
+    public void TestConstructorDouble(decimal d, bool sign, decimal mantissa, int exponent)
     {
         var bd = new BigDecimal(d);
 
@@ -29,7 +29,7 @@ public class BigDecimal_UnitTests
     [InlineData(1.0, 0, true, 1.0, 0)]
     [InlineData(-65473, 3, false, 6.5473, 7)]
     [InlineData(-0.00547, -1, false, 5.47, -4)]
-    public void TestConstructorMantissaExponent(double m, int e, bool sign, double mantissa, int exponent)
+    public void TestConstructorMantissaExponent(decimal m, int e, bool sign, decimal mantissa, int exponent)
     {
         var bd = new BigDecimal(m, e);
 
@@ -43,7 +43,7 @@ public class BigDecimal_UnitTests
     [InlineData(false, 5.2, 4, false, 5.2, 4)]
     [InlineData(true, -5.2, 4, false, 5.2, 4)]
     [InlineData(false, -5.2, 4, true, 5.2, 4)]
-    public void TestConstructorSignMantissaExponent(bool s, double m, int e, bool sign, double mantissa, int exponent)
+    public void TestConstructorSignMantissaExponent(bool s, decimal m, int e, bool sign, decimal mantissa, int exponent)
     {
         var bd = new BigDecimal(s, m, e);
 
@@ -58,7 +58,7 @@ public class BigDecimal_UnitTests
     [InlineData(-3.14, "-3.14E0")]
     [InlineData(50003, "5.0003E4")]
     [InlineData(-0.007, "-7E-3")]
-    public void TestToString(double d, string s)
+    public void TestToString(decimal d, string s)
     {
         new BigDecimal(d).ToString().Should().Be(s);
     }
@@ -67,7 +67,7 @@ public class BigDecimal_UnitTests
     [InlineData(false, 3, 7, false, 30, 6, true)]
     [InlineData(true, 1, 0, false, 1, 0, false)]
     [InlineData(false, 3.54, -7, false, 0.0354, -5, true)]
-    public void TestEquals(bool s1, double m1, int exp1, bool s2, double m2, int exp2, bool isequal)
+    public void TestEquals(bool s1, decimal m1, int exp1, bool s2, decimal m2, int exp2, bool isequal)
     {
         var a = new BigDecimal(s1, m1, exp1);
         var b = new BigDecimal(s2, m2, exp2);
@@ -78,11 +78,11 @@ public class BigDecimal_UnitTests
     public void TestCompare()
     {
         var rand = new Random();
-        var numsDouble = new List<(int, double)>();
+        var numsDouble = new List<(int, decimal)>();
 
         // generate 10k random numbers in the range -150..+150
         for (int i = 1; i <= 10000; i++)
-            numsDouble.Add((i, rand.NextDouble() * 300 - 150));
+            numsDouble.Add((i, (decimal)rand.NextDouble() * 300 - 150));
 
         // copy the same list as BigDecimal
         var numsBigDec = numsDouble.Select(x => (x.Item1, new BigDecimal(x.Item2))).ToList();
@@ -98,7 +98,7 @@ public class BigDecimal_UnitTests
     [InlineData(true, 2, 0, 5, true, 3.2, 1)]
     [InlineData(false, 7, 0, 2, true, 4.9, 1)]
     [InlineData(true, -0.3, 0, 7, true, -0.0002187,0)]
-    public void TestOperatorPower(bool a_sign, double a_m, int a_exp, int n, bool c_sign, double c_m, int c_exp)
+    public void TestOperatorPower(bool a_sign, decimal a_m, int a_exp, int n, bool c_sign, decimal c_m, int c_exp)
     {
         var result = new BigDecimal(a_sign, a_m, a_exp) ^ n;
         result.Should().Be(new BigDecimal(c_sign, c_m, c_exp));
@@ -108,9 +108,9 @@ public class BigDecimal_UnitTests
     [InlineData(true, 2, 3, true, 7, 1, true, 140000, 0)]
     [InlineData(false, 77, -5, true, 13, 1, false, 0.1001, 0)]
     public void TestOperatorMultiplyBigDec(
-        bool a_sign, double a_m, int a_exp,
-        bool b_sign, double b_m, int b_exp,
-        bool c_sign, double c_m, int c_exp)
+        bool a_sign, decimal a_m, int a_exp,
+        bool b_sign, decimal b_m, int b_exp,
+        bool c_sign, decimal c_m, int c_exp)
     {
         var a = new BigDecimal(a_sign, a_m, a_exp);
         var b = new BigDecimal(b_sign, b_m, b_exp);
@@ -124,9 +124,9 @@ public class BigDecimal_UnitTests
     [InlineData(true, 2, 3, 70.0, true, 140000, 0)]
     [InlineData(false, 77, -5, 130.0, false, 0.1001, 0)]
     public void TestOperatorMultiplyDouble(
-        bool a_sign, double a_m, int a_exp,
-        double b,
-        bool c_sign, double c_m, int c_exp)
+        bool a_sign, decimal a_m, int a_exp,
+        decimal b,
+        bool c_sign, decimal c_m, int c_exp)
     {
         var a = new BigDecimal(a_sign, a_m, a_exp);
         var c = new BigDecimal(c_sign, c_m, c_exp);
